+++ tests/test_coupling.py
@@ -6,33 +6,35 @@ from mock import patch
 from scipy.signal import hilbert
 
 from meegsim.coupling import (
-    constant_phase_shift,
+    ppc_constant_phase_shift,
     ppc_von_mises,
     ppc_shifted_copy_with_noise,
+    _get_envelope,
     _get_required_snr,
     _shifted_copy_with_noise,
 )
-from meegsim.utils import get_sfreq, theoretical_plv
+from meegsim.utils import get_sfreq
 
 from utils.prepare import prepare_sinusoid
 
 
-def prepare_inputs():
+def prepare_inputs(sfreq=250, duration=60):
     n_series = 2
-    fs = 1000
-    times = np.arange(0, 1, 1 / fs)
+    times = np.arange(0, duration, 1 / sfreq)
     return n_series, len(times), times
 
 
 @pytest.mark.parametrize(
     "phase_lag", [np.pi / 4, np.pi / 3, np.pi / 2, np.pi, 2 * np.pi]
 )
-def test_constant_phase_shift(phase_lag):
+def test_ppc_constant_phase_shift_same_envelope(phase_lag):
     # Test with a simple sinusoidal waveform
     _, _, times = prepare_inputs()
     waveform = np.sin(2 * np.pi * 10 * times)
 
-    result = constant_phase_shift(waveform, get_sfreq(times), phase_lag)
+    result = ppc_constant_phase_shift(
+        waveform, get_sfreq(times), phase_lag, envelope="same", random_state=1234
+    )
 
     waveform = hilbert(waveform)
     result = hilbert(result)
@@ -41,20 +43,59 @@ def test_constant_phase_shift(phase_lag):
     plv = np.abs(cplv)
     test_angle = np.angle(cplv)
 
-    assert plv >= 0.99, f"Test failed: plv is smaller than 0.99. plv = {plv}"
+    assert plv >= 0.9, f"Expected PLV to be at least 0.9, got {plv}"
+    assert (
+        (np.abs(test_angle) - phase_lag) <= 0.01
+    ), f"Test failed: angle is different from phase_lag. difference = {np.round((np.abs(test_angle) - phase_lag),2)}"
+
+
+@pytest.mark.parametrize(
+    "phase_lag", [np.pi / 4, np.pi / 3, np.pi / 2, np.pi, 2 * np.pi]
+)
+def test_ppc_constant_phase_shift_random_envelope(phase_lag):
+    # Test with a simple sinusoidal waveform
+    _, _, times = prepare_inputs()
+    waveform = np.sin(2 * np.pi * 10 * times)
+
+    result = ppc_constant_phase_shift(
+        waveform,
+        get_sfreq(times),
+        phase_lag,
+        envelope="random",
+        fmin=9.5,
+        fmax=10.5,
+        random_state=1234,
+    )
+
+    waveform = hilbert(waveform)
+    result = hilbert(result)
+
+    cplv = compute_plv(waveform, result, m=1, n=1, plv_type="complex")
+    plv = np.abs(cplv)
+    test_angle = np.angle(cplv)
+
+    assert plv >= 0.9, f"Expected PLV to be at least 0.9, got {plv}"
     assert (
         (np.abs(test_angle) - phase_lag) <= 0.01
     ), f"Test failed: angle is different from phase_lag. difference = {np.round((np.abs(test_angle) - phase_lag),2)}"
 
 
 @pytest.mark.parametrize("m, n", [(2, 1), (3, 1), (5 / 2, 1)])
-def test_constant_phase_shift_harmonics(m, n):
+def test_ppc_constant_phase_shift_harmonics_same_envelope(m, n):
     # Test with different m and n harmonics
     _, _, times = prepare_inputs()
     waveform = np.sin(2 * np.pi * 10 * times)
     phase_lag = np.pi / 3
 
-    result = constant_phase_shift(waveform, get_sfreq(times), phase_lag, m=m, n=n)
+    result = ppc_constant_phase_shift(
+        waveform,
+        get_sfreq(times),
+        phase_lag,
+        envelope="same",
+        m=m,
+        n=n,
+        random_state=1234,
+    )
 
     waveform = hilbert(waveform)
     result = hilbert(result)
@@ -63,21 +104,37 @@ def test_constant_phase_shift_harmonics(m, n):
     plv = np.abs(cplv)
     test_angle = np.angle(cplv)
 
-    assert plv >= 0.9, f"Test failed: plv is smaller than 0.9. plv = {plv}"
+    assert plv >= 0.9, f"Expected PLV to be at least 0.9, got {plv}"
     assert (
         (np.abs(test_angle) - phase_lag) <= 0.1
     ), f"Test failed: angle is different from phase_lag. difference = {np.round((np.abs(test_angle) - phase_lag),2)}"
 
 
-@pytest.mark.parametrize("kappa", [0.001, 0.1, 0.5, 1, 5, 10, 50])
-def test_ppc_von_mises(kappa):
-    # Test kappas that are reliable (more than 0.5)
-    _, _, times = prepare_inputs()
+@pytest.mark.parametrize(
+    "kappa,lo,hi",
+    [
+        (0.001, 0, 0.2),
+        (0.01, 0, 0.2),
+        (0.1, 0.1, 0.4),
+        (0.3, 0.3, 0.7),
+        (0.5, 0.5, 0.9),
+        (1, 0.7, 1),
+        (10, 0.9, 1),
+    ],
+)
+def test_ppc_von_mises_same_envelope_kappa(kappa, lo, hi):
+    _, _, times = prepare_inputs(duration=120)
     waveform = np.sin(2 * np.pi * 10 * times)
-    phase_lag = 0
+    phase_lag = np.pi / 4
 
     result = ppc_von_mises(
-        waveform, get_sfreq(times), phase_lag, kappa=kappa, fmin=8, fmax=12
+        waveform,
+        get_sfreq(times),
+        phase_lag,
+        kappa=kappa,
+        fmin=8,
+        fmax=12,
+        random_state=1234,
     )
 
     waveform = hilbert(waveform)
@@ -85,11 +142,52 @@ def test_ppc_von_mises(kappa):
 
     cplv = compute_plv(waveform, result, m=1, n=1, plv_type="complex")
     plv = np.abs(cplv)
-    plv_theoretical = theoretical_plv(kappa)
+    test_angle = np.abs(np.angle(cplv))
 
-    assert (
-        plv >= plv_theoretical
-    ), f"Test failed: plv is smaller than theoretical. plv = {plv}, plv_theoretical = {plv_theoretical}"
+    # NOTE: lower and upper bounds were selected by simulating multiple time series,
+    # this test should prevent large deviations from the expected result due to
+    # errors in the processing
+    assert lo <= plv <= hi, f"Expected PLV to be between {lo} and {hi}, got {plv}"
+    if kappa >= 0.5:
+        assert np.allclose(test_angle, phase_lag, atol=0.1), (
+            f"Expected the actual phase lag ({test_angle:.2f}) to be within "
+            f"0.1 from the desired one ({phase_lag:.2f})"
+        )
+
+
+@pytest.mark.parametrize("kappa,lo,hi", [(0.01, 0, 0.2), (10, 0.8, 1)])
+def test_ppc_von_mises_random_envelope_kappa(kappa, lo, hi):
+    _, _, times = prepare_inputs(duration=120)
+    waveform = np.sin(2 * np.pi * 10 * times)
+    phase_lag = np.pi / 4
+
+    result = ppc_von_mises(
+        waveform,
+        get_sfreq(times),
+        phase_lag,
+        kappa=kappa,
+        envelope="random",
+        fmin=8,
+        fmax=12,
+        random_state=1234,
+    )
+
+    waveform = hilbert(waveform)
+    result = hilbert(result)
+
+    cplv = compute_plv(waveform, result, m=1, n=1, plv_type="complex")
+    plv = np.abs(cplv)
+    test_angle = np.abs(np.angle(cplv))
+
+    # NOTE: we check only extreme cases here to make sure the random envelope does
+    # not break the result completely. Still, random envelope might decrease PLV a bit,
+    # so the bounds are a bit wider
+    assert lo <= plv <= hi, f"Expected PLV to be between {lo} and {hi}, got {plv}"
+    if kappa >= 0.5:
+        assert np.allclose(np.abs(test_angle), phase_lag, atol=0.1), (
+            f"Expected the actual phase lag ({test_angle:.2f}) to be within "
+            f"0.1 from the desired one ({phase_lag:.2f})"
+        )
 
 
 @pytest.mark.parametrize("m, n", [(2, 1), (3, 1), (5 / 2, 1)])
@@ -97,11 +195,20 @@ def test_ppc_von_mises_harmonics(m, n):
     # Test with different m and n harmonics
     _, _, times = prepare_inputs()
     waveform = np.sin(2 * np.pi * 10 * times)
-    phase_lag = 0
+    phase_lag = np.pi / 4
     kappa = 10
 
     result = ppc_von_mises(
-        waveform, get_sfreq(times), phase_lag, m=m, n=n, kappa=kappa, fmin=8, fmax=12
+        waveform,
+        get_sfreq(times),
+        phase_lag,
+        m=m,
+        n=n,
+        envelope="same",
+        kappa=kappa,
+        fmin=8,
+        fmax=12,
+        random_state=1234,
     )
 
     waveform = hilbert(waveform)
@@ -111,15 +218,20 @@ def test_ppc_von_mises_harmonics(m, n):
     plv = np.abs(cplv)
     test_angle = np.angle(cplv)
 
-    assert plv >= 0.8, f"Test failed: plv is smaller than 0.8. plv = {plv}"
-    assert (
-        (np.abs(test_angle) - phase_lag) <= 0.1
-    ), f"Test failed: angle is different from phase_lag. difference = {np.round((np.abs(test_angle) - phase_lag),2)}"
+    assert 0.7 <= plv, f"Expected PLV to be at least 0.7, got {plv}"
+    assert np.allclose(np.abs(test_angle), phase_lag, atol=0.1), (
+        f"Expected the actual phase lag ({test_angle:.2f}) to be within "
+        f"0.1 from the desired one ({phase_lag:.2f})"
+    )
 
 
 @pytest.mark.parametrize(
     "coupling_fun,params",
     [
+        (
+            ppc_constant_phase_shift,
+            dict(phase_lag=np.pi / 4, envelope="random", fmin=8, fmax=12),
+        ),
         (ppc_von_mises, dict(kappa=1, phase_lag=np.pi / 4, fmin=8, fmax=12)),
         (
             ppc_shifted_copy_with_noise,
@@ -163,6 +275,33 @@ def test_reproducibility_with_random_state(coupling_fun, params):
 )
 def test_get_required_snr(coh, band_limited, expected_snr):
     assert np.isclose(_get_required_snr(coh, band_limited), expected_snr)
+
+
+def test_get_envelope_same():
+    sfreq = 250
+    waveform = prepare_sinusoid(f=10, sfreq=sfreq, duration=60)
+    waveform_amp = np.abs(hilbert(waveform))
+    envelope = _get_envelope(waveform, envelope="same", sfreq=sfreq)
+    assert np.allclose(waveform_amp, envelope), "Expected envelope to match input"
+
+
+def test_get_envelope_random():
+    sfreq = 250
+    fmin, fmax = 8, 12
+    seed = 1234
+    waveform = prepare_sinusoid(f=10, sfreq=sfreq, duration=60)
+    waveform_amp = np.abs(hilbert(waveform))
+    envelope = _get_envelope(
+        waveform,
+        envelope="random",
+        sfreq=sfreq,
+        fmin=fmin,
+        fmax=fmax,
+        random_state=seed,
+    )
+    assert not np.allclose(
+        waveform_amp, envelope
+    ), "Expected envelope not to match input"
 
 
 @pytest.mark.parametrize(